Canvas to Calendar
This program will pull all upcoming assignments due on or after the current day and put them into Google Calendar. This is a work in progress as I would like to make this a web accessible tool that is more flexible and allows for more use cases. I would also like to optimize this as well as enable other people to easily access this tool in the future.